Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240511/합과 평균/sum-and-avg.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-    int a;
-    int b;
+    float a;
+    float b;
 
     cin >>a>>b;
